Creating ingredient widget and refactoring code.
