+++ app/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
             android:name=".ui.DetailStepActivity"
             android:parentActivityName=".ui.DetailActivity" />
 
-        <receiver android:name=".RecipeAppWidget">
+        <receiver android:name=".widget.RecipeAppWidget">
             <intent-filter>
                 <action android:name="android.appwidget.action.APPWIDGET_UPDATE" />
             </intent-filter>
@@ -37,10 +37,10 @@
         </receiver>
 
         <service
-            android:name=".RecipeListWidgetService"
+            android:name=".widget.RecipeListWidgetService"
             android:permission="android.permission.BIND_REMOTEVIEWS" />
         <service
-            android:name=".IngredientListWidgetService"
+            android:name=".widget.IngredientListWidgetService"
             android:permission="android.permission.BIND_REMOTEVIEWS" />
     </application>
