+++ app/src/main/res/layout/fragment_detail_step.xml
@@ -8,15 +8,13 @@
     <com.google.android.exoplayer2.ui.SimpleExoPlayerView
         android:id="@+id/exo_video"
         android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="5" />
+        android:layout_height="wrap_content" />
 
     <TextView
         android:id="@+id/tv_recipe_step_instruction"
         android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="1"
-        android:padding="3dp"
-        android:textAppearance="?android:textAppearanceSmall" />
+        android:layout_height="wrap_content"
+        android:padding="10dp"
+        android:textAppearance="?android:textAppearanceMedium" />
 
 </LinearLayout>
